finish refactor. I really hope this didn't break everything
