+++ .github/actions/fetch-workflow-data/lib/fetching.js
@@ -41,15 +41,17 @@ async function fetchAllWorkflowRuns(github, context, days, cachedRunIds = null, 
   core.info(`[FETCH] days: ${days}, cachedRunIds: ${cachedIds.size}, eventType: ${eventType || 'all'}, branch: ${branch || 'all'}, status: ${status || 'all'}`);
   core.info(`[FETCH] workflowIds: ${workflowIds ? workflowIds.length + ' workflows' : 'all workflows (backward compatibility)'}`);
 
-  const MAX_CONSECUTIVE_CACHED = 100; // Stop after 50 consecutive cached runs
+  const MAX_CONSECUTIVE_CACHED = 100; // Stop after 100 consecutive cached runs
   let totalSkippedOldRuns = 0;
   let totalSkippedCachedRuns = 0;
   let totalAddedNewRuns = 0;
 
+  // Helper function to delay execution
+  const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
+
   // If workflow IDs are provided, fetch runs for each workflow specifically
   if (workflowIds && workflowIds.length > 0) {
     core.info(`[FETCH] Fetching runs for ${workflowIds.length} specific workflows`);
-    const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
 
     for (const workflowId of workflowIds) {
       let consecutiveCachedRuns = 0;
@@ -268,6 +270,204 @@ function groupRunsByName(runs) {
   return grouped;
 }
 
+/**
+ * Helper function to delay execution
+ * @param {number} ms - Milliseconds to delay
+ * @returns {Promise} Promise that resolves after delay
+ */
+const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
+
+/**
+ * Fetches new workflow runs from GitHub API
+ * @param {object} octokit - Octokit client
+ * @param {object} context - GitHub Actions context
+ * @param {number} days - Number of days to look back
+ * @param {Set} cachedRunIds - Set of cached run IDs
+ * @param {Array|null} workflowIds - Array of workflow IDs to fetch, or null for all
+ * @param {string} branch - Branch to filter runs by
+ * @returns {Promise<Array>} Array of new workflow runs
+ */
+async function fetchNewWorkflowRuns(octokit, context, days, cachedRunIds, workflowIds, branch) {
+  // Fetch runs from specified workflows (or all workflows if workflowIds not provided)
+  // Filter at API level: only fetch completed runs on the target branch
+  core.info('Fetching workflow runs...');
+  const allRuns = await fetchAllWorkflowRuns(octokit, context, days, cachedRunIds, workflowIds, '', branch, 'completed');
+  core.info(`[FETCH] Fetched ${allRuns.length} new runs (skipped cached runs during fetch)`);
+
+  // If workflow IDs are provided, we've already fetched all runs for those workflows
+  // Otherwise, for backward compatibility, also fetch scheduled runs separately
+  let newRuns = allRuns;
+  if (!workflowIds || workflowIds.length === 0) {
+    // Wait for 1 second to avoid rate limiting
+    await delay(1000);
+    core.info('[FETCH] Fetching scheduled runs...');
+    const scheduledRuns = await fetchAllWorkflowRuns(octokit, context, days, cachedRunIds, null, 'schedule', branch, 'completed');
+    core.info(`[FETCH] Fetched ${scheduledRuns.length} new scheduled runs (skipped cached runs during fetch)`);
+    // Combine all the results into a single array (already filtered for new runs only)
+    newRuns = [...scheduledRuns, ...allRuns];
+    core.info(`[FETCH] Total new runs fetched: ${newRuns.length} (${allRuns.length} all events + ${scheduledRuns.length} scheduled)`);
+  } else {
+    core.info(`[FETCH] Total new runs fetched: ${newRuns.length} (from ${workflowIds.length} workflows)`);
+  }
+
+  return newRuns;
+}
+
+/**
+ * Merges and deduplicates runs by (run id, attempt) tuple
+ * @param {Array} previousRuns - Array of previous/cached runs
+ * @param {Array} newRuns - Array of newly fetched runs
+ * @param {number} days - Number of days to look back for date filtering
+ * @returns {Array} Array of merged and deduplicated runs
+ */
+function mergeAndDeduplicateRuns(previousRuns, newRuns, days) {
+  // Merge and deduplicate by (run id, attempt) tuple to avoid duplicates
+  // Note: Same run ID with different attempt numbers are different runs
+  core.info(`[MERGE] Merging ${previousRuns.length} cached runs + ${newRuns.length} new runs`);
+  const seen = new Map(); // key: `${run.id}:${run_attempt}`, value: run
+  // Process newRuns first so they naturally take precedence over previousRuns
+  [...newRuns, ...previousRuns].forEach(run => {
+    const runId = run.id;
+    const attempt = run.run_attempt || 1;
+    const key = `${runId}:${attempt}`;
+    // If we already have this exact (run id, attempt) combination, keep the first one encountered.
+    // Since newRuns are processed first, they will naturally take precedence over previousRuns.
+    const existingRun = seen.get(key);
+    if (!existingRun) {
+      seen.set(key, run);
+    }
+  });
+  let mergedRuns = Array.from(seen.values());
+  core.info(`[MERGE] After deduplication: ${mergedRuns.length} unique runs (by run id + attempt)`);
+
+  // Filter by date (branch and status are already filtered at API level)
+  const cutoff = getCutoffDate(days);
+  const beforeFilter = mergedRuns.length;
+  mergedRuns = mergedRuns.filter(run =>
+    new Date(run.created_at) >= cutoff
+  );
+  core.info(`[MERGE] After filtering (date>=${cutoff.toISOString()}): ${mergedRuns.length} runs (removed ${beforeFilter - mergedRuns.length})`);
+  core.info(`[MERGE] Note: branch and status=completed were already filtered at API level`);
+
+  return mergedRuns;
+}
+
+/**
+ * Rechecks workflows for newer run attempts and adds them to the grouped runs
+ * @param {object} octokit - Octokit client
+ * @param {object} context - GitHub Actions context
+ * @param {Map} grouped - Map of workflow names to their runs
+ * @param {Array} workflowConfigs - Array of workflow configuration objects
+ * @param {string} branch - Branch to filter runs by
+ * @param {number} days - Number of days to look back
+ * @param {Set} existingAttemptsSet - Set of existing (run ID, attempt) combinations
+ * @returns {Promise<Array>} Array of new runs with newer attempts to add
+ */
+async function recheckForNewerAttempts(octokit, context, grouped, workflowConfigs, branch, days, existingAttemptsSet) {
+  const { workflowMatchesConfig } = require('./logs');
+  const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
+
+  // Filter workflows to only those matching the configuration (if provided)
+  const workflowsToRecheck = [];
+  if (workflowConfigs.length > 0) {
+    for (const [name, runs] of grouped.entries()) {
+      if (workflowMatchesConfig(name, workflowConfigs)) {
+        workflowsToRecheck.push([name, runs]);
+      }
+    }
+    core.info(`[RECHECK] Filtered to ${workflowsToRecheck.length} workflows matching config (out of ${grouped.size} total)`);
+  } else {
+    // If no configs provided, check all workflows (backward compatibility)
+    workflowsToRecheck.push(...grouped.entries());
+    core.info(`[RECHECK] No workflow_configs provided, checking all ${workflowsToRecheck.length} workflows`);
+  }
+
+  // For each workflow, find the latest run ID (by date, then by attempt) and check for newer attempts
+  const delayBetweenChecks = 100; // Small delay to avoid rate limits
+  let updatedAttempts = 0;
+  const newRunsToAdd = [];
+
+  for (const [workflowName, runs] of workflowsToRecheck) {
+    // Find the latest run for this workflow (on target branch, completed)
+    const latestRuns = runs
+      .filter(r => r.head_branch === branch && r.status === 'completed')
+      .sort((a, b) => {
+        // Sort by date (newest first), then by run_attempt (highest first) as tiebreaker
+        const dateDiff = new Date(b.created_at) - new Date(a.created_at);
+        if (dateDiff !== 0) {
+          return dateDiff;
+        }
+        const attemptA = a.run_attempt || 1;
+        const attemptB = b.run_attempt || 1;
+        return attemptB - attemptA; // Prefer higher attempt number
+      });
+
+    if (latestRuns.length === 0) continue;
+
+    const latestRun = latestRuns[0];
+    const latestRunId = latestRun.id;
+    const currentAttempt = latestRun.run_attempt || 1;
+
+    // Find the highest attempt we have for this run ID (in case there are multiple attempts in the array)
+    let highestAttempt = currentAttempt;
+    for (const run of runs) {
+      if (run.id === latestRunId) {
+        const attempt = run.run_attempt || 1;
+        if (attempt > highestAttempt) {
+          highestAttempt = attempt;
+        }
+      }
+    }
+
+    try {
+      // Fetch the run details directly - GitHub API returns the latest attempt
+      const { data: latestRunData } = await octokit.rest.actions.getWorkflowRun({
+        owner: context.repo.owner,
+        repo: context.repo.repo,
+        run_id: latestRunId
+      });
+
+      const apiAttempt = latestRunData.run_attempt || 1;
+
+      // If API has a higher attempt, add it to mergedRuns (don't replace the old one)
+      if (apiAttempt > highestAttempt) {
+        core.info(`[RECHECK] Found newer attempt for workflow '${workflowName}' run ${latestRunId}: current=${highestAttempt}, API=${apiAttempt}, adding`);
+        // Check if the run is still within our date window and meets other criteria
+        const runDate = new Date(latestRunData.created_at);
+        const cutoff = getCutoffDate(days);
+        if (runDate >= cutoff && latestRunData.head_branch === branch && latestRunData.status === 'completed') {
+          // Check if we already have this attempt (shouldn't happen, but be safe)
+          const alreadyHaveAttempt = existingAttemptsSet.has(`${latestRunId}:${apiAttempt}`);
+          if (!alreadyHaveAttempt) {
+            // Add the new run to the list (don't replace the old one)
+            newRunsToAdd.push(latestRunData);
+            updatedAttempts++;
+          } else {
+            core.info(`[RECHECK] Run ${latestRunId} attempt ${apiAttempt} already in mergedRuns, skipping`);
+          }
+        } else {
+          core.info(`[RECHECK] Run ${latestRunId} attempt ${apiAttempt} is outside date window or doesn't meet criteria, skipping`);
+        }
+      }
+
+      await delay(delayBetweenChecks); // Small delay to avoid rate limits
+    } catch (e) {
+      // Run might not exist anymore, or API error - log and continue
+      if (e.status !== 404) {
+        core.warning(`[RECHECK] Failed to check workflow '${workflowName}' run ${latestRunId} for newer attempts: ${e.message}`);
+      }
+    }
+  }
+
+  if (newRunsToAdd.length > 0) {
+    core.info(`[RECHECK] Found ${updatedAttempts} runs with newer attempts`);
+  } else {
+    core.info(`[RECHECK] No newer attempts found`);
+  }
+
+  return newRunsToAdd;
+}
+
 module.exports = {
   // Constants
   MAX_PAGES,
@@ -277,4 +477,7 @@ module.exports = {
   getCutoffDate,
   fetchAllWorkflowRuns,
   groupRunsByName,
+  fetchNewWorkflowRuns,
+  mergeAndDeduplicateRuns,
+  recheckForNewerAttempts,
 };
